requirements.txt
numpy>=1.21.0
sounddevice>=0.4.5
soundfile>=0.10.3
whispercpp>=0.0.1
openai>=1.0.0
elevenlabs>=0.2.26
python-dotenv>=0.19.0
keyboard>=0.13.5 